LG-683 Allow IdV restart on USPS confirmation (#2541)
**Why**: Because if a user enters an address they can't receive mail at,
or a user moves between requesting a code and returning to the IDP, they
currently have no option to go back and change their address. The commit
allows that by allowing the user to go back and restart IdV.

**How**: This commit creates a new action for cancelling IdV. Profile
deactivation is moved from the phone step into that new action. The "I
can't receive mail at this address" link is changed to reflect that
restarting is a destructive action. After the profile is deactivated,
the user is sent to the beginning of the IdV flow.