Namespace platform authenticator params in analytics controller
**Why**: The analytics controller is meant to function as an general
controller for capture anayltics for the frontend. Prior to this commit
it was primarily handling the platform authenticator analytics event.

**How**: This commit makes the controller more general so that it communicates
that it should be built out to handle more events and moves the platform
auth event into a namespace so that the API can be extended to include
more events. Additionally, care was given to make sure that the current
implementation is supported so that we don't drop events while running
instances running this code alongside instances running the new code.