Guard against nil email in password validator
**Why**: We were seeing `NoMethodErrors` on `email` in production
pointing at this code in the `FormPasswordValidator`:

```ruby
def password_score
  @password_score = ZXCVBN_TESTER.test(
    password,
    ForbiddenPasswords.new(user.email_address.email).call
  )
end
```

When the reset password controller cannot find a user for a given
password reset token, it initializes an empty user with an error.
Once we added the `#email_address` method this started causing problems
because the nil user does not have an email.

This commit fixes the bug and also changes the form to differentiate
between invalid and expired tokens so the user sees the right error
message if their token is invalid.