LG-766 Require MFA after 12 hours for IAL2 and AAL2 (#2639)
* LG-766 Require MFA after 12 hours for IAL2 and AAL2

**Why**: The NIST standards require that MFA be expired after 12 hours.
This commit checks the MFA status during the handoff to the SP and
requires the user to re-auth if their device was remembered and their
cookie is older than 12 hours.

**How**: When a user is MFAed by remember device, flag the session.
Before redirecting to a service provider, check or whether remember
device was used, and if so, if it has expired for the current sp's
configuration. If it has, redirect to the 2FA controller. Also, it is
necessary to prevent the 2FA controller from redirecting if the user is
logged in when the remember device cookie is invalid per the current
SP's configuration.