LG-746 Return 400 error for invalid String params (#2648)
**Why**:
When a controller uses Strong Parameters such as:
`params.require(:user).permit(:email)`, the `user` param is assumed to
be a Hash, but it's easy for a pentester (for example) to set the
`user` param to a String instead, which by default would raise a 500
error because the String class doesn't respond to `permit`.

**How**:
To get around that, we monkey patched the Ruby String class to raise an
instance of `ActionController::ParameterMissing`, which will return a
400 error. 500 errors can potentially page people in the middle of the
night, whereas 400 errors don't.

Note that Devise handles this scenario gracefully in SessionsController,
but because it determines whether or not to call `permit` based on
whether or not the object responds to `permit`, the Devise code will
end up calling `permit` because all Strings respond to it now after our
monkey patch. This is why the `invalid_sign_in_params_spec` had to
change.