Don't tell users to contact us on 2FA key use. (#2708)
**Why**: We don't currently take any action when users write in about
this. This is a very frustrating experience: they are just doing what we
say but we aren't doing anything about it.

**How**: Remove contact us language from email and SMS around personal
key use.

This is just papering over what will still be an urgent problem, so we should reintroduce this language if we later adopt a better strategy for manual handling of account lockouts / takeovers.