Enable configuration for multiple SAML endpoints (#2730)
**Why**: To make it easier to stand up and tear down SAML endpoints by
specifying them in the config rather than using the rotation manager
which was confusing.

This commit also changes the IDP to allow endpoints that use CloudHSM
to be setup in conjuction with endpoints that we don't. If we ever do go
to CloudHSM for signing SAML assertions, this change allows us to move
SPs onto a CloudHSM endpoint gradually instead of all at once.