Use 2L-KMS method to encrypt passwords
**Why**: To move entirely away from using UAKs to encrypt passwords

This commit introduces some code to the user model to migrate users to
the 2L-KMS password digests when they sign in. The UAK code will have to
be kept around for as long as we want to be able to decrypt the
passwords of users who do not sign back in.

The 2L-KMS password verifier is configured to read both 2L-KMS and UAK
passwords. There is a switch that can be used to configure it to write
2L-KMS passwords. The default for production is off. This allows us to
recycle to deploy this code, and then recycle again with that switch
flipped to start writing the 2L-KMS digests when all of the instances in
the pool are able to read them.