Fix issue where PIV/CAC redirects to the account page
**Why**: So that PIV/CAC users can sign in

**How**: The next step method was broken. It was checking for whether a
user had a backup phone setup, but did not take into account other
backup methods. This means that a user with an auth app or backup codes
would be redirected to the recovery page.

The recovery page would see that the users had multiple methods enabled
and then redirect them to the account where they are stranded.

This commit makes some changes to the MFA context decorator to address
the issues above, and also fix the logic around whether a user is MFA
enabled if they have backup codes.