Remove stale password digest rotation
**Why**: The password rotation we are doing currently causes issues when
it is done anywhere besides sign in. This is because rotating the
password digest changes the salt, which Devise uses to verify that the
user in the session is the same user that password authed. This causes
the user be logged out at any point they verify their password outside
of sign in.

We should put something in the rotate the password of users who have
stale digests, but it should only happen on sign in. This commit removes
the rotation entirely to fix issues in int while we figure out a better
approach.