Cleanup 404 rendering around events controller (#2777)
**Why**: To cleanup the code and cover the ActiveModel range error that
happens when device ids are out of range. This commit also uses the
`#find` method on the active record relation to make ActiveRecord raise
when there is an issue finding a device. It rescues the error and
renders a 404.

It also changes the route so that the id is in the URL instead of in a
param. This limits the types of value that the device id can take on.