Add disavowals to events
**Why**: So that users can quickly and easily reset their password
in response to suspicious activity. Additionally, this ties into our
backend analytics pipeline which means we can use users reporting
suspicious activity in this way to build fraud models.

**How**: This commit adds a `disavowed_at` and
`disavowal_token_fingerprint` column to events. When a disavowable event
is created, a token is fingerprinted and saved to the event. That
fingerprint is added to a link and send in the email.

This commit adds a controller for when a user clicks on of those links.
It validates the token, then presents the user with a password reset
form. The user uses that form to reset their password. Once the password
is reset, the disavowal token is invalidated.