Clean up service providers list and add some docs.
**Why**: We probably don't intend to have localhost:// service providers
associated with any deployed environment. In addition, we only intend to
run the sample apps in sandbox environments (identitysandbox.gov) at the
moment, because the sample apps haven't received enough scrutiny to be
deployed in the ATO'd environment.

**How**: Remove all the localhost providers from the production block.
Add some comments about what it means to be a production service
provider. Add logic so that the sample apps are only added when the
server domain is identitysandbox.gov.

Also add some notes about our production service provider policies, such
as that production URIs must use `https://`.