Rename application.yml.example to .default
**Why**: The filename `application.yml.example` is misleading because it
implies that the file contains only example values that are not used
anywhere. In fact, these values are used as defaults in production (with
a deep merge using values from application secrets).

We just never updated the filename to reflect the new reality when we
created the application secrets deep merge mechanism.

**How**: Rename `application.yml.example` to `application.yml.default`
and change all of the references to the file as well.