Add a way to disable recurring jobs using config.
**Why**: We may encounter situations where a recurring job is failing
and we need to rapidly disable it, or where a job is failing and we
don't care so we want health checks to stop failing.

**How**: Add a new Figaro key, `recurring_jobs_disabled_names`. This key
is expected to contain JSON representing an array of string job names.
Any job names present in this list will be ignored when
JobRunner::Runner.add_config is called to add a job with a matching
name. This required refactoring the process to add job configs so that
we have an easy place to add such validation.