Refactor OMB Fitara report to gen S3 bucket name.
**Why**: The existing config methods for the reports had a few
shortcomings. We shouldn't need to set anything in application secrets
in order for this functionality to operate. Any differences between
environments should be encoded in the idp code itself rather than
requiring manual effort from team members anytime a new environment is
created.

In addition, we should use an environment-specific and date-specific
path for the uploaded reports so that we're not overwriting existing
data and so that the IAM permissions can be scoped to prevent
cross-environment interference.

**How**: Add two new Figaro keys, `s3_report_bucket_prefix` and
`s3_reports_enabled`. The latter controls whether we will attempt to
upload to S3 at all (true in production-like envs, false elsewhere). The
former sets the S3 bucket prefix used in all environments. The rest of
the S3 bucket is computed as a function of the environment name, AWS
account ID, and AWS region.