Split job health into normal vs critical checks.
**Why**: For many recurring jobs that we may run, it's not a big deal if
they don't run exactly on schedule or are failing for a while. For
critical jobs, however, we want to know immediately if they fail.

**How**: Add a `health_critical` boolean parameter that can be passed
when defining a new JobConfiguration. If true, the job will be checked
as part of the main app `/api/health` endpoint and will count toward our
top line SLA. If false, the job will only be checked as part of the new
separate `/api/health/jobs` endpoint. This endpoint is expected to be
noisier but purely internal facing.