LG-1789 Update stale password digests as users sign in (#3227)
**Why**: This means that users using digests in legacy formats will have their digests updated to the latest format. This reduces the population of people using legacy formats and makes those legacy formats easier to retire.

This commit uses a devise hook that runs after authentication to rotate the digest. This is necessary because the digest must be rotated between validating the password and saving the user to the session. The user's session depends on the password salt. Redigesting the password changes the password salt and invalidates the user's session if the user has been saved to the session.