Rename service_providers.yml to *.localdev.
**Why**:

The service_providers.yml file contains private, but not secret,
configuration data. This properly belongs in a private github repo, not
in a public open source repo, nor in application secrets.

We're in the process of developing a new strategy for managing
private-but-not-secret config data for the IdP.

**How**:

mv config/service_providers.yml -> config/service_providers.yml.localdev
Create a symlink so the transition is seamless.