Add a hack to work around a fakefs bug.
It turns out that FakeFS does not reset the fake FS state when the block
closes, but reuses it between invocations. This is suprising by itself.

But worse, FakeFS breaks if a symlink exists in the fake filesystem from
a previous run, because it uses FileUtils.ln_s without the force option
set.

We encounter this with the service_providers.yml symlink set up for
local development. So in order to make local tests work, we have to rm
the symlink inside the fake filesystem before cloning.

https://github.com/fakefs/fakefs/issues/343