Symlink logos into assets and symlink SP certs.
To avoid needing to make any changes to the way that logos are served,
we can place the logos into the same rails asset directory that they
were in previously. The deploy/activate script is run before the
deploy/build-post-config script, which is what invokes the asset
pipeline.

Also symlink the SP certs directory, which is used to store SP public
keys.