re-adding CREATE USER statement to grant_readonly_access rake task (#3353)
The IDP was attempting to run a `GRANT SELECT ON ALL TABLES IN SCHEMA public` for a user that didn't exist (ID redacted):

```
** Execute db:grant_readonly_access
rake aborted!
ActiveRecord::StatementInvalid: PG::UndefinedObject: ERROR: role "XXXXXXXXXXXXXXXXXXXXXXXX" does not exist
: GRANT SELECT ON ALL TABLES IN SCHEMA public TO XXXXXXXXXXXXXXXXXXXXXXXX
```

The `CREATE USER` statement was originally in the `readonly_database_user` rake task, which was removed when `grant_readonly_access` was configured. The new task did NOT contain the `CREATE USER` statement, so that has been reinstated here.