LG-2400 Allow any SAML SP to use email addresses as NameIDs
**Why**: There are a lot of SPs using COTS SAML implementations that require them to use the email address as the NameID

We required SPs to be added to a list to use the email address as the NameID in the past. This was to encourage SPs to use the UUID unless they absolutely could not. As we add more SPs this list may become difficult to manage.

This commit allows any SP to use the email address as the NameID and prevents us from having to manage a list of permitted SPs.