Rescue db not found error in migration check. (#3476)
**Why**:

In new personal test environments, the database will not exist the first
time that the pending migration check runs. This causes bootstrapping to
bail out, preventing the database migration process from creating the
database in the first place.

**How**:

Rescue ActiveRecord::NoDatabaseError in
check_for_pending_migrations.rake, as long as the server's domain is not
`login.gov`. This allows the migration process to complete successfully
even when the database doesn't yet exist, which is necessary for
bootstrapping of personal environments.