LG-2715 Add polling to hybrid flow continue step
This commit adds polling to the doc auth hybrid link sent spec. This results in the continue button being hidden. The page automatically advances to the next step when the user is finished on their phone.

This commit includes a feature flag for polling so it can be quickly turned off.