Log time as an iso8601 string, instead of as a float
**Why**: Somehow the event.time property was coming in as
a floating point (and not from any known epoch). As far
as I can tell, that came from the event (which is an
ActiveSupport::Notifications::Event) that was initialized
with a float.

The lograge README has an  example using "Time.now" instead
of referencing "event": https://github.com/roidrage/lograge