LG-2579 Add Support for Multi-Region KMS Encryption (#3812)
Encrypt with KMS in Multiple Regions

Add code to handle the process of doing KMS encryptions across multiple regions, and decrypting appropriately. When data is encrypted through `aws_kms_client`, `multi_region_kms_client` encrypts in every region enumerated in the environment variable `aws_kms_regions`.

In decryption, the client first attempts to decrypt in the current AWS region, or, if that region isn't present in the ciphertext, the first available region. 

Legacy ciphertexts without regions attached are decrypted in the current region.