LG-3207: incorporate selfie in document capture step (#3929)
* LG-3207: Incorporate the selfie upload to the desktop document capture step

As a user, I want to complete all of my doc_auth image uploads in one form so I can proof with a vendor that doesn't allow staggered uploads.

This PR adds the selfie upload to the document_capture_step. It's now one big long form with input fields to allow uploading the front/back of the DL and the selfie image.