Fix a NoMethodError that occured while adding costs in the hybrid flow (#4274)
The old code to add the cost components used `#current_user` which returns nil if there is no user signed in. There is not user in the hybrid flow so this raised `NoMethodError` when `current_user.id` was called.

This commit fixes the bug by reading the user off of the document capture session.

This commit also adjusts the controller tests to fail if there is an assumption the user is logged in.