Use i18n-tasks to audit used JS locale strings
**Why**:

As a developer, I want to be warned when the locale strings in use do not match those provided by the server, so I can have confidence that the texts will be presented correctly to the user.

As a developer, I want to have a well-known inventory of strings in use, so that I can have confidence in future refactoring if and how strings are being used.

As a developer, I want to leverage existing tooling, so that I can feel confidence in knowing that string usage is treated consistently, and so that I don't need to deal with the overhead of maintaining a custom translation string checking utility.

As a user, I want only the necessary translations to be included in the JavaScript assets, so that my pages load faster.