Remove image-preview JavaScript pack
**Why**:

As a user, I don't want to load JavaScript unnecessarily, so that my pages load faster.

As a developer, I want to remove references to classes which are slated for deprecated (alert-error, alert-notice), so that future work is made easier.

As a developer, I want to reduce my number of major dependencies, so that I have a smaller vulnerability attack surface, and fewer APIs to learn.