Add rake task to seed users from a CSV file
Adds the rake task `partners:seed_users` which requires two environment
variables to be set - CSV_FILE and EMAIL_DOMAIN. These are passed to a
`UserSeeder` service object that validates the inputs, ensures that this
task does not run in the upper environments, and proceeds unless it
would cause an existing user to be overwritten. It also then saves a
modified CSV file that includes the backup codes and personal key for
each created user.