Retire the remote settings tooling
The remote setting tooling was added at a time when `service_providers.yml` was tracked in the IdP codebase. At that time it was not possible to update the configs for a service provider without changing the IdP code and deploying a new sha. With identity-idp-config, this is no longer true.

This commit removes the code since now we can modify service provider configs by updating identity-idp-config and recycling to apply the changes.