Make AgencySeeder ignore the abbreviation attribute (#4617)
Temporarily ignore the abbreviation attribute in agencies.yml to allow
the config repo updates to be merged without breaking the IdP while we
work on adding the Partnerships data model to the IdP schema. We will
ultimately add the attribute to the `agencies` table and at that point
remove this change.

Also updates the localdev template YAML file and adds an explicit
fixture file for testing, which is passed to the service object through
dependency injection.