POC: Switch PIV/CAC workflow from 302s to AJAX
In order to mitigate issues with PIN entry delays leading to timeouts
and better manage the PIV/CAC request flow it might be advantageous to
use AJAX requests instead of normal redirects. This commit adds the
necessary JS as well as a JSON response to the `/login/present_piv_cac`
endpoint. It does NOT include tests and is not intended to be a long-
term solution but rather is simply meant to demonstrate feasibility of
this approach.