LG-4175: Control Acuant Web SDK load by device type
**Why**: To avoid script errors in browsers which cannot support Acuant SDK, and to avoid loading Acuant SDK at all in environments where it's not expected to be used.

Prior to these changes, we would always load the Acuant SDK, regardless whether it was planned to be used (i.e. for mobile capture). This was partly toward the expectation that we could use the Acuant SDK to generate metrics for images uploaded in the desktop flow (LG-3436). However, there are also cases where browser requirements of the Acuant SDK script may cause it to fail to load. Specifically, Internet Explorer cannot load the script because it contains modern script syntax (ECMAScript 2015+). This was overlooked in-part because we prevent Internet Explorer users from proceeding to this step of the flow, but importantly we only do this if the service provider requires liveness. Between browser support and the fact that we don't currently use Acuant SDK to generate metrics about uploaded images on desktop, we should prevent loading the script unless it's known to be used.

**Implementation Notes:**

Previously, the implementation of React components consuming from AcuantContext would assume that the Acuant script would eventually be loaded. Thus, the proposed changes here introduce a new context property `isAcuantLoaded` that helps differentiate from existing combinations of properties. "Readiness" now considers whether the user is on a mobile device, and Acuant context is set as immediately ready for desktop users (where `isAcuantLoaded` is set to false).