Allow AgencySeeder to populate abbreviations
Now that the schema has been updated to include the abbreviation column,
we can start to pull them in from the YAML files. The fixture and
localdev files should already include abbreviations. Also add uniqueness
constraint to the DB and the model.

We'll follow this up with a PR to add the null constraint to the column
once all the records are populated.