Add Seeder objects for Agreements data
Resolves LG-3974

Adds a collection of service objects (under the Agreements namespace) to
handle ingesting config YAML files and either creating or updating
records in the database. Leverages inheritance to avoid duplication
across the service objects and requires that a set of private methods be
defined in the child objects.

Due to the fact that we run db:seed in the test environment as well as
in development (and production) this required that a few tests have some
extra deletes added to avoid foreign key constraint violations.