Add Rake task to collect a report of UUIDs for partners
Resolves LG-4250

Adds a partners:get_agency_uuids Rake task and a UuidReporter service
object to find and export a list of agency UUIDs given a list of emails
and a list of ServiceProvider issuers to check. If all of those SPs
belong to the same agency, a user is associated with a given email, and
that user has an Identity with at least one of the SPs, we include the
user's agency UUID in a CSV with their email, otherwise we leave the CSV
blank. Leverages composite indexes on the identities and
agency_identities tables.