Refactor the return to SP url code into a service (#4746)
* Refactor the return to SP url code into a service

**Why**: To make the adding the return to SP event easier by moving the complexity for generating the URL into a service.

This commit also adds some changes to ensure the redirect URI used to return to the SP is in the SP's redirect URI config. This is necessary to prevent an issue where a saml request includes a redirect URI to a host that we don't want to redirect the user to.