Changing our default branch from master to main as part of DE&I efforts on login.gov
As part of our Diversity, Equity and Inclusion efforts on login.gov program, we are intentionally making the default branch of our Identity Provider (IdP) from `master` to `main`. The words “master" and “slave” has been widely used for decades in computing and other technical contexts, as a reference to situations where one process or entity controls another.

If you are still using `master` on your forked version of the IdP, use the following instructions to change to our new `main` default:

- Navigate to the main page of your IdP repository.
- Under your repository name, click  Settings,
- Under "Default branch", to the right of the default branch name, click.
- Use the drop-down, then click a branch name.
- Click Update.
- Read the warning, then click I understand, update the default branch.