LG-4502: Update IAL2 SSN masking and content
**Why**:

As a user, I want to know how my Social Security number will be protected during proofing so that I feel comfortable providing it to verify my identity on login.gov's IAL2 flow.

As a user, I want my Social Security number to be masked by default throughout the process, so that my information is protected if I am using a public computer.