LG-4951: Link relevant help articles in the IAL2 flow
**Why**:

As an end user, I want a way to learn more about the requirements to complete the process on the IAL2 welcome screen, if I have questions about the required documents or information so that I know what I will need to complete identity proofing with login.gov before I get started.

As an end-user, I want a way to learn more about the requirements to complete the phone verification step, so that I can get answers to my questions and know if I will be able to complete identity proofing with login.gov.