Fail JavaScript feature specs if any console messages are logged (#5492)
* Configure global feature spec console log monitoring

**Why**: So that it's less of a mystery when things break

* Refactor message customization as custom class

* Temporarily allow document-capture errors

* Allow skipping expected browser console errors

e.g. 404 page, or specs asserting against an error resposne

* Only assert console logging on supported browser drivers

     NoMethodError:
       undefined method `manage' for #<Capybara::RackTest::Browser:0x000055e95c6122e0>

* Remove redundant visit

Because (a) it's not necessary and (b) it makes console log catching angry because the page will prompt before navigating.

* Fix path for email delete confirmation accessibility

delete_email_path is for the POST submission confirming the deletion, so not something the user would "visit" directly

* Increment reload times for JS-disabled doc polling scenario

It was relying on the page refresh. Better to handle it in the test though, since it matches the test description

* Rename allow_js_error to allow_browser_log

Because it may not be originating from JS, and may not be an error

* Always get logs

Suspected that messages are being held between specs