Ensure SAML POST requests preserve an existing session
Currently, an external POST request to the IdP will not have access to
an existing session since the cookie is blocked due to `SameSite=Lax`.
This results in users with an active session having to sign in again if
sent back to the IdP with a POST SAML request.

However, if the POST request comes _from_ the IdP the session will be
available in that request. Therefore, this commit inserts a new action
within our SAML POST handling that captures the SAML parameters of the
original POST request, populates a form with them (with appropriate CSRF
handling), and then submits that form with JS to create an "internal"
POST request with access to the session. This new action skips session
handling so it does not override an existing IdP session cookie.