Skip select account screen on return visits
Currently, when a user authenticates with an active Login.gov session,
they are taken to an account selection screen as per the OIDC spec. In
order to provide a more seamless "step-up" flow from IAL1 to IAL2, we
are going to only take a user to that screen on their first visit to a
given SP, and skip it on subsequent visits.