Respect the ForceAuthn attribute in SAML AuthnRequests (#5644)
The SAML specification (section 3.4.1) includes an attribute
`ForceAuthn` for the `AuthnRequest` element that can be used by an SP to
require that a user be authenticated again, rather than relying on a
previous security context. This change allows the IdP to respect that
attribute and force authentication when an SP includes
`ForceAuthn='true'` in the SAML request.