Actually fix SAML remote logout
**Why:** The recent updates were not actually working - we were using
the incorrect session key to invalidate the session in redis. This
update stores the session ID on the SP identity record during SAML
authentication (similar to OIDC) so that it can be retrieved during
remote logout. This is actually a safer behavior since it means that a
user will only be logged out if the session they logged in with is still
valid (whereas before they would have been logged out regardless if they
had started a new Login.gov session after authenticating to the SP).

[skip changelog] # covered by entry in #5990