Fix SAML remote logout (#5992)
Correctly invalidate session for SAML Remote Logout (#5990)

**Why:** the way this currently worked led to the user's session being
cleared on subsequent visits, which meant that any authentication
request information was lost. This correctly deletes the session from
the Redis cache. It also refactors out an OutOfBandSessionAccessor
service object for this purpose and adds a feature spec for remote logout.

changelog: Bug Fixes, Authentication, Fix issue with SAML remote logout causing subsequent authentication requests not to redirect correctly

* Actually fix SAML remote logout

**Why:** The recent updates were not actually working - we were using
the incorrect session key to invalidate the session in redis. This
update stores the session ID on the SP identity record during SAML
authentication (similar to OIDC) so that it can be retrieved during
remote logout. This is actually a safer behavior since it means that a
user will only be logged out if the session they logged in with is still
valid (whereas before they would have been logged out regardless if they
had started a new Login.gov session after authenticating to the SP).

Co-authored-by: Zach Margolis <zachary.margolis@gsa.gov>