Improve support for JS string extraction from dynamic imports (#6015)
* Improve support for JS string extraction from dynamic imports

**Why**: To allow translations occurring in dynamically-imported scripts to be loaded.

Currently, string extraction works by adding additional assets associated with a Webpack chunk. However, the runtime script loader responsible for dynamic script imports is not aware of these assets, therefore translations would not be loaded. The workaround is to associate the extracted translations with the closest entrypoint from where the translation is found, so that the translations are loaded alongside the script which would perform the dynamic script import.

* Add changelog

changelog: Internal, Build, Include dynamic JavaScript files for translation string extraction