Confirm personal key for JS-enabled specs
Previously, most of these specs had JS disabled, so it was expected the user would continue to the next step immediately upon clicking personal key "Continue", since we didn't have the modal confirmation in no-JS contexts. But now we're requiring JS for this step, so modal will be shown, and user must enter and confirm their personal key.