Add a flag to not allow a strict IAL2 user to redirect without phone proofing
We may need to make a change in the future to require strict IAL2 users to proof with phone. This commit does some work towards making that possible. It makes changes to the authorization controller to prevent a user from being redirected at strict IAL2 unless they have proofed with phone.

This change is not totally complete, there are changes that will need to go in place to remove the option during proofing.

changelog: Upcoming feature, Proofing, A user will need to proof with a phone before being sent back to the IdP for strict IAL2 proofing