Support IALMAX using the Comparison attribute in SAML (#5652)
**Why:** Several partners have requested the ability to have users sign
in at the maximum level of identity assurance they have obtained without
sending multiple requests to determine whether a user has a verified
credential or not.

The SAML spec does support a `Comparison` attribute for the
`<RequestedAuthnContext>` element that can be set to "exact" (the
default), "minimum", "maximum", or "better". These determine what
authentication context the response should meet relative to the
requested AuthnContext in the SAML request. The specific implementation
of how those are treated is left up to the responder (in this case,
Login.gov).

In this commit, we add the capability for Login.gov to send a user back
with either an auth-only or verified credential to an SP configured to
receive verified attributes when they request the IAL1 AuthnContext with
a Comparison attribute set to "minimum". This does not change the
behavior when the IAL2 AuthnContext is requested or when the SP is not
configured to receive verified attributes.

This also includes more comprehensive feature specs for both the overall
behavior as well as billing records in the `sp_redirect_logs` table.

changelog: Improvements, Authentication, Support IALMAX using the SAML Comparison attribute