Use a depth instead of recursion in the lookup sharerd device users tooling
**Why**: We've had issues with this script running into infinite recursion.

When we wrote this script, we did not have an index on the cookie_uuid column which mean that querying against cookie_uuids was slow and expensive. We now have that index so we can afford to take a more simple approach here.

changelog: Internal, Data requests, The script for looking up users who share devices was changed to avoid infinite recursion.