Update Brakeman to ignore parameter value in render path
Brakeman reports params used in render path. While technically true that it's referenced, it's only used in forwarding "step" parameter for URL creation, used for analytics logging

Next best workaround would be to filter param with an allowlist of steps. This is simple enough, but difficult to maintain, as it would require a developer to remember to add the step to the allowlist. Without automated enforcement checks, this would be far too easy to forget.