LG-5258: Allow users to start over IAL2 verification after clicking Cancel (#6542)
* LG-5258: Allow users to start over IAL2 verification after clicking Cancel

**Why**: As an end user, I expect that when I click Cancel, I will see an option to start over the identity verification process (IAL2), so that I can try to verify my identity again if I do not want to abandon the process completely.

changelog: Improvements, Identity Verification, Revise cancellation screen to allow user to start over

* Remove reference to "Start Over" link

* Remove unnecessary references to "Start Over URL"

No longer managed by the React applications, since the choice to start over is moved into the cancellation experience.

* Rename click_spinner_button to click_spinner_button_and_wait

See: https://github.com/18F/identity-idp/pull/6542#discussion_r911521633
Co-Authored-By: Zach Margolis <zbmargolis@gmail.com>

* Try simplified click_idv_continue

If it turns out to be necessary to pull into viewport, might be good to have in base behavior anyways

Also unsure if we still have multiple instances of "Continue" on the same screen in IdV

Alternatively, update click_spinner_button_and_wait to handle argument as either a locator or as a node

* Update link sent spec to start over via cancel screen

* Wait for spinner button on account page click

* Remove IAL2 cancel spec from ial2_sso_spec

Consolidate cancellation assertions to features/idv/cancel_spec.rb

* Update Brakeman to ignore parameter value in render path

Brakeman reports params used in render path. While technically true that it's referenced, it's only used in forwarding "step" parameter for URL creation, used for analytics logging

Next best workaround would be to filter param with an allowlist of steps. This is simple enough, but difficult to maintain, as it would require a developer to remember to add the step to the allowlist. Without automated enforcement checks, this would be far too easy to forget.

* Normalize YAML

Co-authored-by: Zach Margolis <zbmargolis@gmail.com>