Eliminate IdV async polling delay in test (#6554)
* Try eliminating IdV async polling delay in test

**Why**: It could improve overall test runtime.

In all cases, this polling will wait for a response to be received and handled before issuing a new request, so there should be no risk of overlapping requests overwhelming the local server.

* Add changelog

changelog: Internal, Automated Testing, Improve automated testing performance

* Avoid skipping poll on 0 poll interval

The `|| undefined` logic will be tripped for 0. The intention was to try to avoid scheduling a poll if the result of "Number" parsing was "NaN", since that would be treated the same as "0" in setTimeout, thus potentially causing a DDOS. Instead, check for finiteness of the value when scheduling.